requirements.txt
python-escpos
Pillow
qrcode
requests
flask
flask-restx
